[AFID-363] Refactored to move test data into a fixture.
